Updates time functions in v8::Platform
This adds the following:

1. CurrentClockTimeMilliseconds(), which returns the time in
   milliseconds (int64_t).
2. CurrentClockTimeMillisecondsHighResolution, which returns the time
   with microsecond accuracy (double).

Both are implemented in terms of the existing CurrentClockTimeMillis.

Chrome's implementation of CurrentClockTimeMillis has code to
jitter the results. This code is expensive, and when ms accuracy
is needed (case 1) only necessary a small fraction of the time.
Adding CurrentClockTimeMilliseconds allows chrome to optimize this
case.


Bug: chromium:1414615
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4246257
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#85869}

+++ src/objects/js-objects.h
@@ -1114,8 +1114,8 @@ class JSDate : public TorqueGeneratedJSDate<JSDate, JSObject> {
   static V8_WARN_UNUSED_RESULT MaybeHandle<JSDate> New(
       Handle<JSFunction> constructor, Handle<JSReceiver> new_target, double tv);
 
-  // Returns the time value (UTC) identifying the current time.
-  static double CurrentTimeValue(Isolate* isolate);
+  // Returns the time value (UTC) identifying the current time in milliseconds.
+  static int64_t CurrentTimeValue(Isolate* isolate);
 
   // Returns the date field with the specified index.
   // See FieldIndex for the list of date fields.
